+++ src/main/java/tech/ydb/io/r2dbc/statement/DDLYdbStatement.java
@@ -16,12 +16,10 @@
 
 package tech.ydb.io.r2dbc.statement;
 
-import java.sql.SQLException;
 
 import io.r2dbc.spi.Result;
 import io.r2dbc.spi.Statement;
 import org.reactivestreams.Publisher;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import tech.ydb.io.r2dbc.query.YdbQuery;
 import tech.ydb.io.r2dbc.state.YdbConnectionState;
@@ -33,39 +31,35 @@ public class DDLYdbStatement extends YdbStatement {
     public DDLYdbStatement(YdbQuery query, YdbConnectionState ydbConnectionState) {
         super(query, ydbConnectionState);
     }
+    @Override
+    public Statement add() {
+        return this;
+    }
 
     @Override
     public Statement bind(int i, Object o) {
-        return null;
+        return this;
     }
 
     @Override
     public Statement bind(String s, Object o) {
-        return null;
+        return this;
     }
 
     @Override
     public Statement bindNull(int i, Class<?> aClass) {
-        return null;
+        return this;
     }
 
     @Override
     public Statement bindNull(String s, Class<?> aClass) {
-        return null;
+        return this;
     }
 
     @Override
     public Publisher<? extends Result> execute() {
         try {
-            return Flux.fromIterable(bindings)
-                    .flatMap(binding -> {
-                        try {
-                            return connectionState.executeDataQuery(query.getYqlQuery(bindings.getCurrent()),
-                                            bindings.getCurrent());
-                        } catch (SQLException e) {
-                            return Mono.error(e);
-                        }
-                    });
+            return connectionState.executeSchemaQuery(query.getYqlQuery(bindings.getCurrent()), bindings.getCurrent());
         } catch (Exception e) {
             return Mono.error(e);
         }
